lab3 almost done

+++ src/FordFulkerson/FordFulkerson.java
@@ -12,8 +12,8 @@ import java.util.HashMap;
 public class FordFulkerson {
     Graph<Integer> network = new Graph<Integer>(new MatrixStorage<>());
     Graph<Integer> residual = new Graph<Integer>(new MatrixStorage<>());
-    HashMap<Vertex, HashMap<Vertex, Float>> flows;
-    HashMap<Vertex, HashMap<Vertex, Float>> cf;
+    HashMap<Vertex, HashMap<Vertex, Integer>> flows;
+    HashMap<Vertex, HashMap<Vertex, Integer>> cf;
 
     public void run(int sourceVal, int targetVal) throws IOException {
         importGraphs();
@@ -26,7 +26,7 @@ public class FordFulkerson {
 
         Edge[] mainPath = residual.findPath(source, target);
 
-        float minFlow = getMinAdditionalFlow(mainPath);
+        int minFlow = getMinAdditionalFlow(mainPath);
 
         while(minFlow > 0) {
             for (Edge e : mainPath) {
@@ -36,14 +36,14 @@ public class FordFulkerson {
         }
     }
 
-    public float getMinAdditionalFlow(Edge[] path) {
-        Float min = null;
+    public int getMinAdditionalFlow(Edge[] path) {
+        Integer min = null;
 
         for (Edge e : path) {
             if (null == min) {
                 min = e.getWeight() - getFlow(e.getStart(), e.getEnd());
             } else {
-                float additionalFlow = e.getWeight() - getFlow(e.getStart(), e.getEnd());
+                int additionalFlow = e.getWeight() - getFlow(e.getStart(), e.getEnd());
                 if (additionalFlow < min) {
                     min = additionalFlow;
                 }
@@ -57,13 +57,13 @@ public class FordFulkerson {
 
     }
 
-    private float getFlow(Vertex start, Vertex end) {
-        HashMap<Vertex, Float> map = flows.get(start);
+    private int getFlow(Vertex start, Vertex end) {
+        HashMap<Vertex, Integer> map = flows.get(start);
         if (null == map) {
             return 0;
         }
 
-        Float flow = map.get(end);
+        Integer flow = map.get(end);
 
         if (null == flow) {
             return 0;
